requirements.txt
# Retail Sales Data Warehouse - Requirements

# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Database Connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Jupyter and Analytics
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Environment Management
python-dotenv>=0.21.0

# Testing (optional, for development)
pytest>=7.2.0
pytest-cov>=4.0.0

# Data Quality
great-expectations>=0.15.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.7
